+++ lbz/misc.py
@@ -5,6 +5,7 @@ Misc Helpers of Lambda Framework.
 import copy
 import logging
 import logging.handlers
+import warnings
 from collections.abc import MutableMapping
 from functools import wraps
 from os import environ
@@ -132,3 +133,29 @@ def deep_update(dict_to_update: dict, update_data: dict) -> None:
 
 def is_in_debug_mode() -> bool:
     return environ.get("LBZ_DEBUG_MODE", "").lower() == "true"
+
+
+def deprecated(message: str, version: str) -> Callable:
+    """
+    This is a decorator which can be used to mark functions as deprecated.
+
+    It will result in a warning being emitted when the function is used.
+    """
+
+    def decorator(func: Callable) -> Callable:
+        warnings.warn(
+            (
+                f"Call to deprecated function {func.__name__}. {message}. "
+                f"Function will be removed in {version}."
+            ),
+            category=DeprecationWarning,
+            stacklevel=2,
+        )
+
+        @wraps(func)
+        def wrapped(*args: Any, **kwargs: Any) -> Any:
+            return func(*args, **kwargs)
+
+        return wrapped
+
+    return decorator